feat: implement accept and decline invitations
